requirements.txt
Django==1.11
PyMySQL==0.7.11
requests==2.18.4
django-crontab==0.7.1
djangorestframework==3.7.7
django-redis==4.8.0
